requirements.txt
numpy~=1.21.2
tensorflow~=2.6.0
tqdm~=4.36.1
mtcnn==0.1.1
opencv-python~=4.5.4.58
pandas~=1.3.0
matplotlib~=3.1.3
albumentations==0.4.6
pytest~=6.2.5
Pillow~=8.4.0
efficientnet